+++ GeneralsMD/Code/GameEngine/Source/GameLogic/Object/Collide/CrateCollide/SabotageInternetCenterCrateCollide.cpp
@@ -18,7 +18,8 @@
 
 ////////////////////////////////////////////////////////////////////////////////
 //																																						//
-//  (c) 2001-2003 Electronic Arts Inc.																				//
+//  (c) 2001-2003 Electronic Arts Inc.
+//  //
 //																																						//
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -26,14 +27,14 @@
 //
 // FILE: SabotageInternetCenterCrateCollide.cpp
 // Author: Kris Morness, July 2003
-// Desc:   A crate (actually a saboteur - mobile crate) that temporarily disables an internet center
+// Desc:   A crate (actually a saboteur - mobile crate) that temporarily
+// disables an internet center
 //
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 
-
-
-// INCLUDES ///////////////////////////////////////////////////////////////////////////////////////
-#include "PreRTS.h"	// This must go first in EVERY cpp file in the GameEngine
+// INCLUDES
+// ///////////////////////////////////////////////////////////////////////////////////////
+#include "PreRTS.h" // This must go first in EVERY cpp file in the GameEngine
 
 #include "Common/GameAudio.h"
 #include "Common/MiscAudio.h"
@@ -46,7 +47,7 @@
 #include "GameClient/Drawable.h"
 #include "GameClient/Eva.h"
 #include "GameClient/GameText.h"
-#include "GameClient/InGameUI.h"  // useful for printing quick debug strings when we need to
+#include "GameClient/InGameUI.h" // useful for printing quick debug strings when we need to
 
 #include "GameLogic/ExperienceTracker.h"
 #include "GameLogic/Object.h"
@@ -62,173 +63,159 @@
 #include "GameLogic/Module/SabotageInternetCenterCrateCollide.h"
 #include "GameLogic/Module/SpyVisionUpdate.h"
 
-
-
 //-------------------------------------------------------------------------------------------------
 //-------------------------------------------------------------------------------------------------
-SabotageInternetCenterCrateCollide::SabotageInternetCenterCrateCollide( Thing *thing, const ModuleData* moduleData ) : CrateCollide( thing, moduleData )
-{
-}
+SabotageInternetCenterCrateCollide::SabotageInternetCenterCrateCollide(
+    Thing *thing, const ModuleData *moduleData)
+    : CrateCollide(thing, moduleData) {}
 
 //-------------------------------------------------------------------------------------------------
 //-------------------------------------------------------------------------------------------------
-SabotageInternetCenterCrateCollide::~SabotageInternetCenterCrateCollide( void )
-{
-}
+SabotageInternetCenterCrateCollide::~SabotageInternetCenterCrateCollide(void) {}
 
 //-------------------------------------------------------------------------------------------------
 //-------------------------------------------------------------------------------------------------
-Bool SabotageInternetCenterCrateCollide::isValidToExecute( const Object *other ) const
-{
-	if( !CrateCollide::isValidToExecute(other) )
-	{
-		//Extend functionality.
-		return FALSE;
-	}
+Bool SabotageInternetCenterCrateCollide::isValidToExecute(
+    const Object *other) const {
+  if (!CrateCollide::isValidToExecute(other)) {
+    // Extend functionality.
+    return FALSE;
+  }
 
-	if( other->isEffectivelyDead() )
-	{
-		//Can't sabotage dead structures
-		return FALSE;
-	}
+  if (other->isEffectivelyDead()) {
+    // Can't sabotage dead structures
+    return FALSE;
+  }
 
-	if( !other->isKindOf( KINDOF_FS_INTERNET_CENTER ) )
-	{
-		//We can only sabotage supply dropzones.
-		return FALSE;
-	}
+  if (!other->isKindOf(KINDOF_FS_INTERNET_CENTER)) {
+    // We can only sabotage supply dropzones.
+    return FALSE;
+  }
 
 #if !RETAIL_COMPATIBLE_CRC
-	if (other->getStatusBits().testForAny(MAKE_OBJECT_STATUS_MASK2(OBJECT_STATUS_UNDER_CONSTRUCTION, OBJECT_STATUS_SOLD)))
-	{
-		// TheSuperHackers @bugfix Stubbjax 03/08/2025 Can't enter something being sold or under construction.
-		return FALSE;
-	}
+  if (other->getStatusBits().testForAny(MAKE_OBJECT_STATUS_MASK2(
+          OBJECT_STATUS_UNDER_CONSTRUCTION, OBJECT_STATUS_SOLD))) {
+    // TheSuperHackers @bugfix Stubbjax 03/08/2025 Can't enter something being
+    // sold or under construction.
+    return FALSE;
+  }
 #endif
 
-	Relationship r = getObject()->getRelationship( other );
-	if( r != ENEMIES )
-	{
-		//Can only sabotage enemy buildings.
-		return FALSE;
-	}
+  Relationship r = getObject()->getRelationship(other);
+  if (r != ENEMIES) {
+    // Can only sabotage enemy buildings.
+    return FALSE;
+  }
 
-	return TRUE;
+  return TRUE;
 }
 
-static void disableHacker( Object *obj, void *userData )
-{
-	UnsignedInt frame = (UnsignedInt)userData;
-	if( obj )
-	{
-		obj->setDisabledUntil( DISABLED_HACKED, frame );
-	}
+static void disableHacker(Object *obj, void *userData) {
+  UnsignedInt frame = (UnsignedInt)(uintptr_t)userData;
+  if (obj) {
+    obj->setDisabledUntil(DISABLED_HACKED, frame);
+  }
 }
 
-static void disableInternetCenterSpyVision( Object *obj, void *userData )
-{
-	if( obj && obj->isKindOf( KINDOF_FS_INTERNET_CENTER ) )
-	{
-		UnsignedInt frame = (UnsignedInt)userData;
+static void disableInternetCenterSpyVision(Object *obj, void *userData) {
+  if (obj && obj->isKindOf(KINDOF_FS_INTERNET_CENTER)) {
+    UnsignedInt frame = (UnsignedInt)(uintptr_t)userData;
 
-		//Loop through all it's SpyVisionUpdates() and wake them all up so they can be shut down. This is weird because
-		//it's one of the few update modules that is actually properly sleepified.
-		for( BehaviorModule** u = obj->getBehaviorModules(); *u; ++u )
-		{
-			SpyVisionUpdate *svUpdate = (*u)->getSpyVisionUpdate();
-			if( svUpdate )
-			{
-				//Turn off the vision temporarily. When it recovers from being disabled, the
-				//timer will need to start over from scratch so it won't come on right away unless
-				//it's a permanent spy vision.
-				svUpdate->setDisabledUntilFrame( frame );
-			}
-		}
-	}
+    // Loop through all it's SpyVisionUpdates() and wake them all up so they can
+    // be shut down. This is weird because it's one of the few update modules
+    // that is actually properly sleepified.
+    for (BehaviorModule **u = obj->getBehaviorModules(); *u; ++u) {
+      SpyVisionUpdate *svUpdate = (*u)->getSpyVisionUpdate();
+      if (svUpdate) {
+        // Turn off the vision temporarily. When it recovers from being
+        // disabled, the timer will need to start over from scratch so it won't
+        // come on right away unless it's a permanent spy vision.
+        svUpdate->setDisabledUntilFrame(frame);
+      }
+    }
+  }
 }
 
 //-------------------------------------------------------------------------------------------------
 //-------------------------------------------------------------------------------------------------
-Bool SabotageInternetCenterCrateCollide::executeCrateBehavior( Object *other )
-{
-	//Check to make sure that the other object is also the goal object in the AIUpdateInterface
-	//in order to prevent an unintentional conversion simply by having the terrorist walk too close
-	//to it.
-	//Assume ai is valid because CrateCollide::isValidToExecute(other) checks it.
-	Object *obj = getObject();
-	AIUpdateInterface* ai = obj->getAIUpdateInterface();
-	if (ai && ai->getGoalObject() != other)
-	{
-		return false;
-	}
+Bool SabotageInternetCenterCrateCollide::executeCrateBehavior(Object *other) {
+  // Check to make sure that the other object is also the goal object in the
+  // AIUpdateInterface in order to prevent an unintentional conversion simply by
+  // having the terrorist walk too close to it. Assume ai is valid because
+  // CrateCollide::isValidToExecute(other) checks it.
+  Object *obj = getObject();
+  AIUpdateInterface *ai = obj->getAIUpdateInterface();
+  if (ai && ai->getGoalObject() != other) {
+    return false;
+  }
 
-	TheRadar->tryInfiltrationEvent( other );
+  TheRadar->tryInfiltrationEvent(other);
 
-  doSabotageFeedbackFX( other, CrateCollide::SAB_VICTIM_INTERNET_CENTER );
+  doSabotageFeedbackFX(other, CrateCollide::SAB_VICTIM_INTERNET_CENTER);
 
-	if( other->isLocallyViewed() )
-	{
-		TheEva->setShouldPlay( EVA_BuildingSabotaged );
-	}
+  if (other->isLocallyViewed()) {
+    TheEva->setShouldPlay(EVA_BuildingSabotaged);
+  }
 
-	//Loop through every internet center to temporarily disable the spy vision upgrades.
-	UnsignedInt frame = TheGameLogic->getFrame() + getSabotageInternetCenterCrateCollideModuleData()->m_sabotageFrames;
+  // Loop through every internet center to temporarily disable the spy vision
+  // upgrades.
+  UnsignedInt frame =
+      TheGameLogic->getFrame() +
+      getSabotageInternetCenterCrateCollideModuleData()->m_sabotageFrames;
 
-	//Disable all internet center spy visions (they stack) without visually disabling the other centers.
-	//Kris: Note -- Design has changed that we only can have one center at a time... logically, this code
-	//doesn't need to change.
-	// This loop goes before the Disabled_Hacked one since that will use the normal disabled code with its cool timers.
-	// This loop is for the other centers, but it hits the main one too.
-	other->getControllingPlayer()->iterateObjects( disableInternetCenterSpyVision, (void*)frame );
+  // Disable all internet center spy visions (they stack) without visually
+  // disabling the other centers. Kris: Note -- Design has changed that we only
+  // can have one center at a time... logically, this code doesn't need to
+  // change.
+  //  This loop goes before the Disabled_Hacked one since that will use the
+  //  normal disabled code with its cool timers. This loop is for the other
+  //  centers, but it hits the main one too.
+  other->getControllingPlayer()->iterateObjects(disableInternetCenterSpyVision,
+                                                (void *)frame);
 
-	//Disable the internet center. Note this is purely fluff... because the spy vision update will still run even
-	//though we are disabling it. We have to disable the spyvision updates manually because other centers need to
-	//be turned off too but without visually disabling them. Yikes!
-	other->setDisabledUntil( DISABLED_HACKED, frame );
+  // Disable the internet center. Note this is purely fluff... because the spy
+  // vision update will still run even though we are disabling it. We have to
+  // disable the spyvision updates manually because other centers need to be
+  // turned off too but without visually disabling them. Yikes!
+  other->setDisabledUntil(DISABLED_HACKED, frame);
 
-	//Disable all the hackers inside.
-	ContainModuleInterface *contain = other->getContain();
-	contain->iterateContained( disableHacker, (void*)frame, FALSE );
+  // Disable all the hackers inside.
+  ContainModuleInterface *contain = other->getContain();
+  contain->iterateContained(disableHacker, (void *)frame, FALSE);
 
-	return TRUE;
+  return TRUE;
 }
 
 // ------------------------------------------------------------------------------------------------
 /** CRC */
 // ------------------------------------------------------------------------------------------------
-void SabotageInternetCenterCrateCollide::crc( Xfer *xfer )
-{
+void SabotageInternetCenterCrateCollide::crc(Xfer *xfer) {
 
-	// extend base class
-	CrateCollide::crc( xfer );
-
+  // extend base class
+  CrateCollide::crc(xfer);
 }
 
 // ------------------------------------------------------------------------------------------------
 /** Xfer method
-	* Version Info:
-	* 1: Initial version */
+ * Version Info:
+ * 1: Initial version */
 // ------------------------------------------------------------------------------------------------
-void SabotageInternetCenterCrateCollide::xfer( Xfer *xfer )
-{
+void SabotageInternetCenterCrateCollide::xfer(Xfer *xfer) {
 
-	// version
-	XferVersion currentVersion = 1;
-	XferVersion version = currentVersion;
-	xfer->xferVersion( &version, currentVersion );
+  // version
+  XferVersion currentVersion = 1;
+  XferVersion version = currentVersion;
+  xfer->xferVersion(&version, currentVersion);
 
-	// extend base class
-	CrateCollide::xfer( xfer );
-
+  // extend base class
+  CrateCollide::xfer(xfer);
 }
 
 // ------------------------------------------------------------------------------------------------
 /** Load post process */
 // ------------------------------------------------------------------------------------------------
-void SabotageInternetCenterCrateCollide::loadPostProcess( void )
-{
+void SabotageInternetCenterCrateCollide::loadPostProcess(void) {
 
-	// extend base class
-	CrateCollide::loadPostProcess();
-
+  // extend base class
+  CrateCollide::loadPostProcess();
 }
